fixes de casteo y variables erroneas

+++ Sprint 3/BackendGeems/BackendGeems/Infraestructure/PagoRepo.cs
@@ -180,19 +180,19 @@ namespace BackendGeems.Infraestructure
             }
             else if (ingresoMensual <= limite2)
             {
-                impuesto = (int)((ingresoMensual - limite1) * porcentaje1);
+                impuesto = ((ingresoMensual - limite1) * porcentaje1);
             }
             else if (ingresoMensual <= limite3)
             {
-                impuesto = (int)((limite2 - limite1) * porcentaje1 + (ingresoMensual - limite2) * porcentaje2);
+                impuesto = ((limite2 - limite1) * porcentaje1 + (ingresoMensual - limite2) * porcentaje2);
             }
             else if (ingresoMensual <= limite4)
             {
-                impuesto = (int)((limite2 - limite1) * porcentaje1 + (limite3 - limite2) * porcentaje2 + (ingresoMensual - limite3) * porcentaje3);
+                impuesto = ((limite2 - limite1) * porcentaje1 + (limite3 - limite2) * porcentaje2 + (ingresoMensual - limite3) * porcentaje3);
             }
             else
             {
-                impuesto = (int)((limite2 - limite1) * porcentaje1 + (limite3 - limite2) * porcentaje2 + (limite4 - limite3) * porcentaje3 + (ingresoMensual - limite4) * porcentaje4);
+                impuesto = ((limite2 - limite1) * porcentaje1 + (limite3 - limite2) * porcentaje2 + (limite4 - limite3) * porcentaje3 + (ingresoMensual - limite4) * porcentaje4);
             }
 
             return impuesto;
